LayoutTests/http/tests/media/video-cookie-expected.txt

EVENT(canplay)
Tests that the media player will send the relevant cookies when requesting the media file.
